Added some pizaz to the Changes Page
The lastseen is only a temporary solution at the moment. I also will
need to convert all dates to UTC dates, store the spotifyID in the
database so that I can make links to play the tracks from the web app on
mobile and come up with a better way to display changes to the user.