Changing to Built-In Auth Provider for Spotify
I am creating a custom OAuth plugin for Spotify by leveraging the
Microsoft.Owin crap but I just added that project to this solution and
for the first time it did not die and my web app is currently working.
Therfore I am saving before it all breaks in a moment when I enable the
custom spotify auth I created.